requirements.txt
streamlit>=1.29.0
fastapi>=0.108.0
uvicorn>=0.25.0
temporalio>=1.4.0
couchbase>=4.3.0
boto3>=1.34.0
pydantic>=2.5.3
pandas>=2.1.4
plotly>=5.18.0
python-dotenv>=1.0.0
httpx>=0.25.2
networkx>=3.5
pytest>=7.4
pytest-asyncio>=0.23.0
voyageai>=0.3.5
groq